Move away all behavior related to legacy message
